+++ .gitattributes
@@ -1,27 +1,27 @@
 # ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
 
-*.snap                             	linguist-generated
-/.devcontainer.json                	linguist-generated
-/.eslintrc.json                    	linguist-generated
-/.gitattributes                    	linguist-generated
-/.github/pull_request_template.md  	linguist-generated
-/.github/workflows/auto-approve.yml	linguist-generated
-/.github/workflows/build.yml       	linguist-generated
-/.github/workflows/release.yml     	linguist-generated
-/.github/workflows/semantic-pr.yml 	linguist-generated
-/.github/workflows/stale.yml       	linguist-generated
-/.github/workflows/upgrade-main.yml	linguist-generated
-/.gitignore                        	linguist-generated
-/.gitpod.yml                       	linguist-generated
-/.markdownlint.json                	linguist-generated
-/.mergify.yml                      	linguist-generated
-/.npmignore                        	linguist-generated
-/.projen/**                        	linguist-generated
-/.projen/deps.json                 	linguist-generated
-/.projen/tasks.json                	linguist-generated
-/.vscode/launch.json               	linguist-generated
-/LICENSE                           	linguist-generated
-/package.json                      	linguist-generated
-/projen.bash                       	linguist-generated
-/tsconfig.dev.json                 	linguist-generated
-/yarn.lock                         	linguist-generated
+*.snap                                  	linguist-generated
+/.devcontainer.json                     	linguist-generated
+/.eslintrc.json                         	linguist-generated
+/.gitattributes                         	linguist-generated
+/.github/pull_request_template.md       	linguist-generated
+/.github/workflows/auto-approve.yml     	linguist-generated
+/.github/workflows/build.yml            	linguist-generated
+/.github/workflows/pull-request-lint.yml	linguist-generated
+/.github/workflows/release.yml          	linguist-generated
+/.github/workflows/stale.yml            	linguist-generated
+/.github/workflows/upgrade-main.yml     	linguist-generated
+/.gitignore                             	linguist-generated
+/.gitpod.yml                            	linguist-generated
+/.markdownlint.json                     	linguist-generated
+/.mergify.yml                           	linguist-generated
+/.npmignore                             	linguist-generated
+/.projen/**                             	linguist-generated
+/.projen/deps.json                      	linguist-generated
+/.projen/tasks.json                     	linguist-generated
+/.vscode/launch.json                    	linguist-generated
+/LICENSE                                	linguist-generated
+/package.json                           	linguist-generated
+/projen.bash                            	linguist-generated
+/tsconfig.dev.json                      	linguist-generated
+/yarn.lock                              	linguist-generated
