+++ .github/workflows/pull-request-lint.yml
@@ -1,6 +1,6 @@
 # ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
 
-name: semantic-pr
+name: pull-request-lint
 on:
   pull_request_target:
     types:
@@ -24,14 +24,5 @@ jobs:
           types: |-
             feat
             fix
-            docs
-            style
-            refactor
-            perf
-            test
-            build
-            ci
             chore
-            revert
-            release
           requireScope: false
