+++ src/github/index.ts
@@ -8,7 +8,7 @@ export * from './mergify';
 export * from './auto-merge';
 export * from './auto-approve';
 export * from './stale';
-export * from './semantic-pr';
+export * from './pull-request-lint';
 
 export * from './task-workflow';
